+++ notion_database/__init__.py
@@ -42,8 +42,8 @@ from notion_database.models.properties import PropertySchema, PropertyValue
 from notion_database.models.rich_text import RichText
 from notion_database.models.sorts import Sort
 
-NOTION_VERSION = "2022-06-28"
-__version__ = "2.0.dev1"
+NOTION_VERSION = "2026-03-11"
+__version__ = "2.0.0rc1"
 
 __all__ = [
     # Client